Backend/requirements.txt
django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.3
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.1